Equipment Window Bug Fix
Problem - The Equipment Window was not displaying which weapons were currently being equipped by the player.

Solution - EquipmentWindowUI Script is now attached to the "Equipment Weapons" UI component rather than the "Equipment Window" UI component. This will need to be fixed later so it works on the "Equipment Window" UI component but at least I now know what the issue is.