Enemy AI State Bug Fix
Problem: Enemy is leaving animations before they should be able to. For example, when the enemy is parried, they do not play the entire parried animation and instead move to another state.

Solution: Make a check to see if the enemy is currently interacting at the beginning of each state,