Player/Enemy attack bug fix
Problem: Player and Enemy could not both use the OpenDamageCollider CloseDamageCollider functions during attack animation events.

Solution: At first I was going to make an abstract CharacterWeoponSlotManager which the other two classes inherit from. This class would contain an abstract DamageCollider, while the other two class would contain overrides of these functions. Then I realised the reason this bug existed was because both functions in the EnemyWeaponSlotManager class were spelled wrong lol. This is why it could not be shared in the animation event with the player functions.